(refactor) Refactor actions
