+++ client/src/actions/mealsActions.js
@@ -1,4 +1,5 @@
 import axios from 'axios';
+import _ from 'lodash';
 
 import { attachTokenToTheHeader } from '../utils/RequestHelper';
 
@@ -7,6 +8,7 @@ export const FETCH_MEALS = 'FETCH_MEALS';
 export const FETCH_MEALDETAIL = 'FETCH_MEALDETAIL';
 export const FETCH_MEALS_BY_CHEF = 'FETCH_MEALS_BY_CHEF';
 export const GET_NEAR_BY_MEALS = 'GET_NEAR_BY_MEALS';
+export const FETCH_TODAYS_MEALS = 'FETCH_TODAYS_MEALS';
 
 export function createMeal(values, cb) {
   const headers = attachTokenToTheHeader();
@@ -52,4 +54,25 @@ export function getNearbyMeals(zipcode) {
     type: GET_NEAR_BY_MEALS,
     payload: request
   }
+}
+
+export function fetchTodaysMeals() {
+  let today = new Date();
+  let year = today.getFullYear().toString();
+  let month = (today.getMonth() + 1).toString().length === 1 ? '0' + (today.getMonth() + 1).toString() : (today.getMonth() + 1).toString();
+  let day = today.getDate().toString().length === 1 ? '0' + today.getDate().toString() : today.getDate().toString();
+  let date = year + '-' + month + '-' + day;
+
+  let request = axios.get('/api/meals')
+    .then(function(meals) {
+      console.log(meals);
+      return _.filter(meals.data, (meal) => {
+        return meal.deliveryDateTime.substr(0, 10) === date;
+      });
+    });
+
+  return {
+    type: FETCH_TODAYS_MEALS,
+    payload: request
+  };
 }
