+++ client/src/index.js
@@ -34,7 +34,6 @@ ReactDOM.render(
             <PrivateRoute exact user='chef' path="/post-new-meal" component={NewMealForm} />
             <Route exact path="/log-in-form" component={LogInForm} />
             <Route exact path="/chef-profile" component={ChefProfile}/>
-            <Route exact path="/" component={Homepage} />
           </Switch>
         </div>
       </BrowserRouter>
